requirements.txt
Django>=4.2
mysqlclient
djangorestframework
python-dotenv